Put route files back to a sane state
If you ask me, pretending that there are routes for the ‘web’ (read: accessed by browsers) and that there are other, different routes for APIs is pure bullshit. The Web is, by definition, supposed to be hypermedia. So there is just one, and only one, valid type of routes.

This commit removes one of these two route files and rename the remaining one in a neutral way: `routes.php`

It also cleans the `RouteServiceProvider` by removing the mess induced by the theory of the two sets of routes.